+++ sdk/python/pulumi_azure/compute/availability_set.py
@@ -5,7 +5,7 @@
 import warnings
 import pulumi
 import pulumi.runtime
-from typing import Any, Dict, List, Mapping, Optional, Tuple, Union
+from typing import Any, Mapping, Optional, Sequence, Union
 from .. import _utilities, _tables
 
 __all__ = ['AvailabilitySet']
@@ -18,8 +18,8 @@ class AvailabilitySet(pulumi.CustomResource):
                  location: Optional[pulumi.Input[str]] = None,
                  managed: Optional[pulumi.Input[bool]] = None,
                  name: Optional[pulumi.Input[str]] = None,
-                 platform_fault_domain_count: Optional[pulumi.Input[float]] = None,
-                 platform_update_domain_count: Optional[pulumi.Input[float]] = None,
+                 platform_fault_domain_count: Optional[pulumi.Input[int]] = None,
+                 platform_update_domain_count: Optional[pulumi.Input[int]] = None,
                  proximity_placement_group_id: Optional[pulumi.Input[str]] = None,
                  resource_group_name: Optional[pulumi.Input[str]] = None,
                  tags: Optional[pulumi.Input[Mapping[str, pulumi.Input[str]]]] = None,
@@ -49,8 +49,8 @@ class AvailabilitySet(pulumi.CustomResource):
         :param pulumi.Input[str] location: Specifies the supported Azure location where the resource exists. Changing this forces a new resource to be created.
         :param pulumi.Input[bool] managed: Specifies whether the availability set is managed or not. Possible values are `true` (to specify aligned) or `false` (to specify classic). Default is `true`.
         :param pulumi.Input[str] name: Specifies the name of the availability set. Changing this forces a new resource to be created.
-        :param pulumi.Input[float] platform_fault_domain_count: Specifies the number of fault domains that are used. Defaults to `3`.
-        :param pulumi.Input[float] platform_update_domain_count: Specifies the number of update domains that are used. Defaults to `5`.
+        :param pulumi.Input[int] platform_fault_domain_count: Specifies the number of fault domains that are used. Defaults to `3`.
+        :param pulumi.Input[int] platform_update_domain_count: Specifies the number of update domains that are used. Defaults to `5`.
         :param pulumi.Input[str] proximity_placement_group_id: The ID of the Proximity Placement Group to which this Virtual Machine should be assigned. Changing this forces a new resource to be created
         :param pulumi.Input[str] resource_group_name: The name of the resource group in which to create the availability set. Changing this forces a new resource to be created.
         :param pulumi.Input[Mapping[str, pulumi.Input[str]]] tags: A mapping of tags to assign to the resource.
@@ -95,8 +95,8 @@ class AvailabilitySet(pulumi.CustomResource):
             location: Optional[pulumi.Input[str]] = None,
             managed: Optional[pulumi.Input[bool]] = None,
             name: Optional[pulumi.Input[str]] = None,
-            platform_fault_domain_count: Optional[pulumi.Input[float]] = None,
-            platform_update_domain_count: Optional[pulumi.Input[float]] = None,
+            platform_fault_domain_count: Optional[pulumi.Input[int]] = None,
+            platform_update_domain_count: Optional[pulumi.Input[int]] = None,
             proximity_placement_group_id: Optional[pulumi.Input[str]] = None,
             resource_group_name: Optional[pulumi.Input[str]] = None,
             tags: Optional[pulumi.Input[Mapping[str, pulumi.Input[str]]]] = None) -> 'AvailabilitySet':
@@ -110,8 +110,8 @@ class AvailabilitySet(pulumi.CustomResource):
         :param pulumi.Input[str] location: Specifies the supported Azure location where the resource exists. Changing this forces a new resource to be created.
         :param pulumi.Input[bool] managed: Specifies whether the availability set is managed or not. Possible values are `true` (to specify aligned) or `false` (to specify classic). Default is `true`.
         :param pulumi.Input[str] name: Specifies the name of the availability set. Changing this forces a new resource to be created.
-        :param pulumi.Input[float] platform_fault_domain_count: Specifies the number of fault domains that are used. Defaults to `3`.
-        :param pulumi.Input[float] platform_update_domain_count: Specifies the number of update domains that are used. Defaults to `5`.
+        :param pulumi.Input[int] platform_fault_domain_count: Specifies the number of fault domains that are used. Defaults to `3`.
+        :param pulumi.Input[int] platform_update_domain_count: Specifies the number of update domains that are used. Defaults to `5`.
         :param pulumi.Input[str] proximity_placement_group_id: The ID of the Proximity Placement Group to which this Virtual Machine should be assigned. Changing this forces a new resource to be created
         :param pulumi.Input[str] resource_group_name: The name of the resource group in which to create the availability set. Changing this forces a new resource to be created.
         :param pulumi.Input[Mapping[str, pulumi.Input[str]]] tags: A mapping of tags to assign to the resource.
@@ -156,7 +156,7 @@ class AvailabilitySet(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="platformFaultDomainCount")
-    def platform_fault_domain_count(self) -> pulumi.Output[Optional[float]]:
+    def platform_fault_domain_count(self) -> pulumi.Output[Optional[int]]:
         """
         Specifies the number of fault domains that are used. Defaults to `3`.
         """
@@ -164,7 +164,7 @@ class AvailabilitySet(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="platformUpdateDomainCount")
-    def platform_update_domain_count(self) -> pulumi.Output[Optional[float]]:
+    def platform_update_domain_count(self) -> pulumi.Output[Optional[int]]:
         """
         Specifies the number of update domains that are used. Defaults to `5`.
         """
